backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-jose[cryptography]
bcrypt
python-multipart
python-dotenv